IOS-4118 Fix
